Refactor DTO package structure
- Organized `DTOs` into sub-packages for improved modularity and readability
- Created specific packages: `AuthorDto`, `BookDto`, `CustomerDto`, `SaleDto`.
