+++ bookstore/src/main/java/com/bookstore/bookstore/service/AuthorService.java
@@ -1,7 +1,6 @@
 package com.bookstore.bookstore.service;
 
-import com.bookstore.bookstore.dto.AuthorsDto;
-import com.bookstore.bookstore.entity.Author;
+import com.bookstore.bookstore.dto.AuthorDto.AuthorsDto;
 import com.bookstore.bookstore.mapper.AuthorMapper;
 import com.bookstore.bookstore.repository.AuthorRepository;
 import lombok.extern.slf4j.Slf4j;
