+++ bookstore/src/main/java/com/bookstore/bookstore/service/BookService.java
@@ -1,9 +1,9 @@
 package com.bookstore.bookstore.service;
 
-import com.bookstore.bookstore.dto.BookDetailsDto;
-import com.bookstore.bookstore.dto.BookDto;
-import com.bookstore.bookstore.dto.CreateBookDto;
-import com.bookstore.bookstore.dto.UpdateBookDto;
+import com.bookstore.bookstore.dto.BookDto.BookDetailsDto;
+import com.bookstore.bookstore.dto.BookDto.BookDto;
+import com.bookstore.bookstore.dto.BookDto.CreateBookDto;
+import com.bookstore.bookstore.dto.BookDto.UpdateBookDto;
 import com.bookstore.bookstore.exception.BookAlreadyExistsException;
 import com.bookstore.bookstore.entity.Book;
 import com.bookstore.bookstore.exception.BookNotFoundException;
